+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 📋 Changelog for "LOOPIS Content"
 
-## 0.31 (2026-02-26)
+## 0.3.2 (2026-03-09)
+- Tags instead of categories for FAQ posts
+- Updated null and save function for custom fields
+
+## 0.3.1 (2026-02-26)
 - Some fixes for CPTs
 - Translated some comments
 - Cleanup